backend/requirements.txt
Flask==3.0.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
supabase>=2.3.0
postgrest>=0.16.0
gunicorn>=21.2.0
openai>=1.0.0
